feat:
Scanner complete. A shopping cart of keys can be generated, they will slide across the screen and then the user can click the appropriate input to slide them along. They are deleted after they slide off of the screen. Currently the scanner is generating a preview cart, that will be removed once the game manager is created.